modify the autoformer interface  and docs
