--- examples/nas/oneshot/autoformer/README_zh_CN.md
@@ -1,1 +1,0 @@
-[文档](https://nni.readthedocs.io/zh/latest/NAS/Autoformer.html)
