+++ examples/nas/oneshot/autoformer/supernet_engine.py
@@ -1,0 +1,67 @@
+from timm.utils.model import unwrap_model
+import torch
+from lib import utils
+import random
+
+
+def sample_configs(choices):
+
+    config = {}
+    dimensions = ['mlp_ratio', 'num_heads']
+    depth = random.choice(choices['depth'])
+    for dimension in dimensions:
+        config[dimension] = [random.choice(choices[dimension]) for _ in range(depth)]
+
+    config['embed_dim'] = [random.choice(choices['embed_dim'])]*depth
+
+    config['layer_num'] = depth
+    return config
+
+
+@torch.no_grad()
+def evaluate(data_loader, model, device, amp=True, choices=None, mode='super', retrain_config=None):
+    criterion = torch.nn.CrossEntropyLoss()
+
+    metric_logger = utils.MetricLogger(delimiter="  ")
+    header = 'Test:'
+
+    # switch to evaluation mode
+    model.eval()
+    if mode == 'super':
+        config = sample_configs(choices=choices)
+        model_module = unwrap_model(model)
+        model_module.set_sample_config(config=config)
+    else:
+        config = retrain_config
+        model_module = unwrap_model(model)
+        model_module.set_sample_config(config=config)
+
+
+    print("sampled model config: {}".format(config))
+    parameters = model_module.get_sampled_params_numel(config)
+    print("sampled model parameters: {}".format(parameters))
+
+    for images, target in metric_logger.log_every(data_loader, 10, header):
+        images = images.to(device, non_blocking=True)
+        target = target.to(device, non_blocking=True)
+        # compute output
+        if amp:
+            with torch.cuda.amp.autocast():
+                output = model(images)
+                loss = criterion(output, target)
+        else:
+            output = model(images)
+            loss = criterion(output, target)
+
+        acc1, acc5 = accuracy(output, target, topk=(1, 5))
+
+        batch_size = images.shape[0]
+        metric_logger.update(loss=loss.item())
+        metric_logger.meters['acc1'].update(acc1.item(), n=batch_size)
+        metric_logger.meters['acc5'].update(acc5.item(), n=batch_size)
+    # gather the stats from all processes
+    metric_logger.synchronize_between_processes()
+    print('* Acc@1 {top1.global_avg:.3f} Acc@5 {top5.global_avg:.3f} loss {losses.global_avg:.3f}'
+          .format(top1=metric_logger.acc1, top5=metric_logger.acc5, losses=metric_logger.loss))
+
+    return {k: meter.global_avg for k, meter in metric_logger.meters.items()}
