--- nni/algorithms/nas/pytorch/autoformer/__init__.py
@@ -1,1 +1,0 @@
-from .trainer import AFSupernetTrainer
