+++ nni/retiarii/oneshot/pytorch/__init__.py
@@ -6,3 +6,4 @@ from .enas import EnasTrainer
 from .proxyless import ProxylessTrainer
 from .random import SinglePathTrainer, RandomTrainer
 from .utils import replace_input_choice, replace_layer_choice
+from .autoformer import AFSupernetTrainer
